* compiler: avoid creating target output folder twice (should speed-up compile time) * compiler: avoid integer overflow in depslib * compiler: avoid checking for filetype twice - compiler: some code optimisations and more debugging output in case of failures during command line generation
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@8082 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/compilergcc/depslib/src/depslib.h
@@ -22,10 +22,10 @@ extern "C" {
 /* This structure is returned by depsGetStats() */
 struct depsStats
 {
-	int scanned; /* the number of files scanned for #include */
-	int cache_used; /* This many files were not scanned for #include because
+	long scanned; /* the number of files scanned for #include */
+	long cache_used; /* This many files were not scanned for #include because
 					   up-to-date entries in the cache were found */
-	int cache_updated; /* This many files had entries in the cache updated */
+	long cache_updated; /* This many files had entries in the cache updated */
 };
 
 typedef struct _depsRef *depsRef;
